ES-2356 Fix PSR-1 and PSR-4 violations, move generated and vendor files out to top level directories
