+++ lib/class_aliases.php
@@ -1,0 +1,3 @@
+<?php
+
+class_alias(\LightStepBase\GlobalLightStep::class, '\LightStep');
