+++ phpunit.xml.dist
@@ -4,6 +4,7 @@
     processIsolation="false"
     stopOnFailure="false"
     syntaxCheck="false"
+    bootstrap="vendor/autoload.php"
     >
     <php>
         <ini name="memory_limit" value="-1"/>
