+++ thrift/Thrift/Exception/TApplicationException.php
