+++ thrift/Thrift/Protocol/TBinaryProtocolAccelerated.php
