+++ thrift/Thrift/Protocol/TProtocolDecorator.php
