+++ thrift/Thrift/StringFunc/Core.php
