+++ thrift/Thrift/StringFunc/Mbstring.php
