+++ thrift/Thrift/StringFunc/TStringFunc.php
